src/utils/helper.py
import os

def read_binary_file(file_path):
    """
    This function reads a binary file and returns its content.
    """
    with open(file_path, 'rb') as file:
        return file.read()

def calculate_size(file_path):
    """
    This function calculates the size of a file in bytes.
    """
    return os.path.getsize(file_path)

def find_dependencies(file_path):
    """
    This function finds the dependencies of a binary file.
    For simplicity, let's assume it returns an empty list for now.
    """
    return []